fix：レフトサイドバーにモーダル機能を集約
