+++ src/components/sidebars/LeftSideBar/index.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { Planet, Edit, LogIn, Settings } from 'iconoir-react';
 import Image from 'next/image';
 
@@ -9,23 +9,63 @@ import {ProfileButton} from '@/components/buttons/ProfileButton';
 import { ActionButton } from "@/components/buttons/ActionButton";
 import { PrimaryButton } from "@/components/buttons/PrimaryButton";
 
-import SettingsModal from "@/components/modals/SettingsModal"
+import { MemoModal } from '@/components/modals/MemoModal';
+import { AccountModal } from '@/components/modals/AccountModal';
+import SettingsModal from "@/components/modals/SettingsModal";
 
-interface LeftSideBarProps {
-  isLogin: boolean; 
-  onMemoModal: () => void;
-  onOpenModal: () => void;
-}
+import { createClient } from '@supabase/supabase-js';
 
-export function LeftSideBar({ isLogin, onMemoModal, onOpenModal }: LeftSideBarProps) {
+const supabase = createClient(
+  process.env.NEXT_PUBLIC_SUPABASE_URL!,
+  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
+);
+
+
+export function LeftSideBar() {
   const [isSettingsModalOpen, setIsSettingsModalOpen] = useState(false);
+  const [isMemoModalOpen, setIsMemoModalOpen] = useState(false);
+  const [isAccountModalOpen, setIsAccountModalOpen] = useState(false);
+  const [isLogin, setIsLogin] = useState(false);
+
+  useEffect(() => {
+    // セッションをチェックしてログイン状態を設定
+    const checkSession = async () => {
+      const { data: { session }, error } = await supabase.auth.getSession();
+      if (error) {
+        console.error('Failed to get session:', error.message);
+      } else if (session) {
+        setIsLogin(true);
+      }
+    };
+    checkSession();
+
+    // 認証状態の変化を監視
+    const { data: authListener } = supabase.auth.onAuthStateChange(
+      (event) => {
+        if (event === 'SIGNED_IN') {
+          setIsLogin(true);
+        } else if (event === 'SIGNED_OUT') {
+          setIsLogin(false);
+        }
+      }
+    );
+
+    // クリーンアップ関数
+    return () => {
+      authListener.subscription.unsubscribe();
+    };
+  }, []);
+
 
   const buttonTitle = isLogin ? "メモする" : "ログイン";
   const buttonIcon = isLogin ? Edit : LogIn;
-  const buttonAction = isLogin ? onMemoModal : onOpenModal;
+  const buttonAction = isLogin ? () => setIsMemoModalOpen(true) : () => setIsAccountModalOpen(true);
 
   const openSettingsModal = () => setIsSettingsModalOpen(true);
   const closeSettingsModal = () => setIsSettingsModalOpen(false);
+
+  const closeMemoModal = () => setIsMemoModalOpen(false);
+  const closeAccountModal = () => setIsAccountModalOpen(false);
 
   return (
     <div className="pt-8">
@@ -40,11 +80,16 @@ export function LeftSideBar({ isLogin, onMemoModal, onOpenModal }: LeftSideBarPr
         </div>
         <ProfileButton title="HIMAZIN"  path = "kitune" hideTextOnSmallScreen={true}/>
         <ActionButton title="みんな" path="/" icon={Planet} />
-        <ActionButton title="設定" path="/setting" icon={Settings} />
         <PrimaryButton title={buttonTitle} icon={buttonIcon} onClick={buttonAction} hideTextOnSmallScreen={true}/>
-        <PrimaryButton title="設定" icon={Planet} onClick={openSettingsModal}/>
+        <PrimaryButton title="設定" icon={Settings} onClick={openSettingsModal}/>
       </div>
       <SettingsModal isOpen={isSettingsModalOpen} onClose={closeSettingsModal} />
+      <MemoModal isOpen={isMemoModalOpen} onClose={closeMemoModal} />
+      <AccountModal
+        isOpen={isAccountModalOpen}
+        onClose={closeAccountModal}
+        onLogin={() => setIsLogin(true)}
+      />
     </div>
     
   );
